Be specific with Python version in Dockerfile
CloudPebble currently uses Python 2.7.8 in production. We know that changes in Python patch versions can change behaviour which effects CloudPebble (e.g. SNI/SSL stuff introduced in 2.7.9) so we should be specific about the version of Python we develop with the Dockerfile.

A small side effect is that 2.7.8 ships with an older version of pip, which needs to be upgraded to support `pip install --no-cache-dir`.